# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Busbud Suggestion API",
        "description": "This is a challenge API for Busbud. Users can search destinations via location name, latitude and longitude",
        "termsOfService": "https://www.busbud.com/en/about/tos",
        "contact": {
            "email": "atayurtmert@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://busbud-city-app.herokuapp.com/",
            "description": "API Host"
        }
    ],
    "paths": {
        "/suggestions": {
            "get": {
                "tags": [
                    "suggestions"
                ],
                "summary": "Get Suggestions",
                "description": "Get suggestions for given search param.",
                "operationId": "getSuggestionsByParams",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Location Name",
                        "required": true,
                        "style": "form",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "latitude",
                        "in": "query",
                        "description": "Location Latitude",
                        "style": "form",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "description": "Location Longitude",
                        "style": "form",
                        "schema": {
                            "type": "number",
                            "format": "float"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "400": {
                        "description": "Invalid name value"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "City": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64"
                    },
                    "name": {
                        "type": "string"
                    },
                    "lat": {
                        "type": "number",
                        "format": "float"
                    },
                    "long": {
                        "type": "number",
                        "format": "float"
                    },
                    "score": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "City"
                }
            }
        }
    },
    "security": [
        []
    ]
}